# docs/openapi.yaml
openapi: 3.0.3
info:
  title: 0dteTrader API
  version: 0.1.0
servers:
  - url: /v1
paths:
...
  /orders/preview:
    post:
      tags: [trading]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OrderRequest' }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/OrderPreview' } } } }
  /orders:
    post:
      tags: [trading]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: Idempotency-Key, in: header, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OrderRequest' }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/OrderResult' } } } }
        '403': { description: Trading disabled (kill switch) }
    get:
      tags: [trading]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/OrderResult' } } } } }
  /orders/{orderId}:
    delete:
      tags: [trading]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: orderId, in: path, required: true, schema: { type: string } }
      responses:
        '204': { description: Cancelled }
  /positions:
    get:
      tags: [trading]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Position' } } } } }
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  schemas:
...
    OrderRequest:
      type: object
      required: [underlying, assetClass, side, quantity, orderType, selection]
      properties:
        underlying: { type: string }
        assetClass: { type: string, enum: [option, future] }
        side: { type: string, enum: [buy, sell] }
        quantity: { type: integer, minimum: 1 }
        orderType: { type: string, enum: [mid, market] }
        selection:
          type: object
          required: [mode]
          properties:
            mode: { type: string, enum: [auto_otm, explicit] }
            optionType: { type: string, enum: [call, put] }
            expiration: { type: string, format: date }
            strike: { type: number }
            contractSymbol: { type: string }
    OrderPreview:
      type: object
      properties:
        resolved:
          type: object
          properties:
            contractSymbol: { type: string }
            price: { type: number }
            estBuyingPower: { type: number }
        warnings: { type: array, items: { type: string } }
    OrderResult:
      type: object
      properties:
        orderId: { type: string }
        status: { type: string, enum: [submitted, filled, partially_filled, cancelled, rejected] }
        contractSymbol: { type: string }
        side: { type: string }
        quantity: { type: integer }
        orderType: { type: string }
        limitPrice: { type: number }
        filledPrice: { type: number }
        timestamp: { type: string, format: date-time }
    Position:
      type: object
      properties:
        symbol: { type: string }
        assetClass: { type: string }
        quantity: { type: integer }
        avgPrice: { type: number }
        markPrice: { type: number }
        unrealizedPnl: { type: number }